Улучшена структура интерфейсной библиотеки
