+++ bitrix/ryir/Components/ryir/interface.input.checkbox.multiple/class.php
@@ -1,9 +1,9 @@
 <?php
 
 use Ryir\Core\Component\Base;
-use Ryir\Core\Traits\ComponentTrait;
+use Ryir\Core\Traits\InterfaceComponentTrait;
 
 class InterfaceInputCheckboxMultiple extends Base
 {
-    use ComponentTrait;
+    use InterfaceComponentTrait;
 }
