+++ bitrix/ryir/Components/ryir/interface.input.checkbox/templates/template.php
@@ -1,0 +1,10 @@
+<div 
+<?isset($result['div_class']) ? print 'class="' . $result['div_class'] . '"' : '';?>
+>
+<button
+type="<?=$result['type'];?>"
+<?isset($result['button_class']) ? print 'class="' . $result['button_class'] . '"' : '';?>
+>
+<?isset($result['default']) ? print $result['default'] : '';?>
+</button>
+</div>
