+++ bitrix/ryir/Components/ryir/interface.input.text.multiple/class.php
@@ -1,9 +1,9 @@
 <?php
 
 use Ryir\Core\Component\Base;
-use Ryir\Core\Traits\ComponentTrait;
+use Ryir\Core\Traits\InterfaceComponentTrait;
 
 class InterfaceInputExtMultiple extends Base
 {
-    use ComponentTrait;
+    use InterfaceComponentTrait;
 }
