+++ bitrix/ryir/Components/ryir/interface.select.multiple/class.php
@@ -1,9 +1,9 @@
 <?php
 
 use Ryir\Core\Component\Base;
-use Ryir\Core\Traits\ComponentTrait;
+use Ryir\Core\Traits\InterfaceComponentTrait;
 
 class InterfaceSelectMultiple extends Base
 {
-    use ComponentTrait;
+    use InterfaceComponentTrait;
 }
