+++ bitrix/ryir/Core/Validator.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Ryir\Core;
+
+class Validator
+{
+//     $valid = new Validator('minLength', 5);
+// $valid->exec($value) : boolean result
+// или , когда сложная валидация
+
+// $valid = new Validator
+//     (
+//     'chain',
+//      true,
+//      [
+//         new Validator('minLength', 5),
+//         new Validator('regexp', '/^[A-Za-z0-9]{0,}$/'),
+//         new Validator('email')
+//      ]
+//     );
+
+// $valid->exec($value);
+
+public function __construct()
+{
+    
+}
+
+}
