+++ bitrix/templates/alice/components/navbar.php
@@ -1,5 +1,3 @@
-
-
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
   <div class="container-fluid">
     <a class="navbar-brand"></a>
@@ -28,4 +26,3 @@
     </div>
   </div>
 </nav>
-
